centralize unique named iterator
