+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/headers/HeadersHash.scala
@@ -65,67 +65,9 @@ private[http] final class HeadersHash extends mutable.HashMap[String, Header.Roo
     array.iterator
   }
 
-  def uniqueNamesIterator: Iterator[String] = new Iterator[String] {
-    private[this] val it = copiedEntitiesIterator
-    private[this] var current: Iterator[String] = Iterator.empty
-
-    private[this] def collectUnique(from: Header): Iterator[String] = new Iterator[String] {
-
-      // We need to keep `_next` current so that we can reliably determine `.hasNext`.
-      // That means that if `_next` is not null, it is, in fact, the next value.
-      private[this] var nextUnique = from
-
-      // We keep a set of observed names so that we don't accidentally return
-      // duplicates. Another way to have done this would be to simply add all names
-      // to a set and return an iterator over that set, but this way is more lazy.
-      private[this] var seen = Set.empty + from.name
-
-      def hasNext: Boolean = nextUnique != null
-
-      def next(): String = {
-        if (!hasNext) throw new NoSuchElementException
-
-        val result = nextUnique.name
-        // We then need to determine the next `_next` so we can
-        // again have a sane `.hasNext`.
-        prepareNext()
-        result
-      }
-
-      @tailrec
-      private[this] def prepareNext(): Unit = {
-        nextUnique = nextUnique.next
-        if (nextUnique != null) {
-          val prevSize = seen.size
-          seen = seen + nextUnique.name
-          if (seen.size == prevSize) {
-            // already observed. Try again.
-            prepareNext()
-          }
-        }
-      }
-    }
-
-    def hasNext: Boolean =
-      it.hasNext || current.hasNext
-
-    def next(): String = {
-      if (current.isEmpty) {
-        val hs = it.next()
-        if (hs.next == null) {
-          // A shortcut to see if we have only a single entry for this name
-          hs.name
-        } else {
-          // slow path: we need to make a sub-iterator
-          current = collectUnique(hs)
-          current.next()
-        }
-      } else {
-        current.next()
-      }
-    }
-  }
-
+  def uniqueNamesIterator: Iterator[String] =
+    Header.uniqueNames(copiedEntitiesIterator)
+    
   def flattenIterator: Iterator[(String, String)] =
     flattenedNameValueIterator.map(nv => (nv.name, nv.value))
