+++ finagle-benchmark/src/main/scala/com/twitter/finagle/http/HeaderMapBenchmark.scala
@@ -119,5 +119,5 @@ class HashBackedHeaderMapBenchmark extends HeaderMapBenchmark {
 }
 
 class JTMapBackedHeaderMapBenchmark extends HeaderMapBenchmark {
-  protected def newMap(): HeaderMap = new JTreeMapBackedHeaderMap()
+  protected def newMap(): HeaderMap = JTreeMapBackedHeaderMap()
 }
